requirement.txt
colorama
tqdm
inquirer
pytest
wheel
setuptools
Pillow
